+++ src/main/java/org/frc5687/steamworks/protobot/commands/actions/dustpan/MoveDustpan.java
@@ -1,0 +1,7 @@
+package org.frc5687.steamworks.protobot.commands.actions.dustpan;
+
+/**
+ * Created by Baxter on 3/11/2017.
+ */
+public class MoveDustpan {
+}
